Use separate ViewCopy header for benchmarks to avoid gtest dependency
