+++ core/perf_test/PerfTest_ViewCopy_c8.cpp
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include <PerfTest_ViewCopy.hpp>
+#include <Benchmark_ViewCopy.hpp>
 
 // DeepCopy Performance for LayoutLeft to LayoutRight
 namespace Test {
