+++ core/perf_test/PerfTest_ViewCopy_d8.cpp
@@ -42,7 +42,7 @@
 //@HEADER
 */
 
-#include <PerfTest_ViewCopy.hpp>
+#include <Benchmark_ViewCopy.hpp>
 
 // DeepCopy Performance for LayoutRight to LayoutLeft
 namespace Test {
